+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     - add icons/.desktop files to linux install
 - Fix next/prev pen shortcut keys ('[' and ']').
 - Sync colour picking between palette editor and main editor window.
+- Stash spritesheet info in PNG metadata to ease reloading
 
 ## v0.2.1 (Jan 2021)